// book-network-ui/src/app/services/model/bookRequest.ts
/**
 * OpenApi specification - Jayarrakesh
 *
 * Contact: prabakars2003@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface BookRequest { 
    id?: number;
    title: string;
    authorName: string;
    isbn: string;
    sysnopsis: string;
    shareable?: boolean;
}